requirements.txt
python-binance==1.0.19
pandas==2.1.0
python-dotenv==1.0.0 